+++ scripts/VVS/resource.hpp
